% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsgtheme_reset.R
\name{rsgtheme_reset}
\alias{rsgtheme_reset}
\title{MReset theme}
\usage{
rsgtheme_reset()
}
\description{
This function uses the ggthemr package to reset the document theme to the
default colors from ggplot2
}
