% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_rsg_d.R
\name{scale_color_rsg_d}
\alias{scale_color_rsg_d}
\title{Function to scale colors to the RSG theme}
\usage{
scale_color_rsg_d()
}
\description{
This function uses the ggthemr package to create a function to scale colored plots to RSG colors.
It will use the colors loaded in the rsgtheme function.
}
